+++ benches/bench-memchr.rs
@@ -1,7 +1,7 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 mod barrier;
-use barrier::memory_barrier;
+use barrier::cache_line_flush;
 
 #[inline(never)]
 fn process_std_memchr(texts: &[&str], pat_byte: u8) -> usize {
@@ -140,27 +140,12 @@ fn process_memx_memchr_sse2(texts: &[&str], pat_byte: u8) -> usize {
     found
 }
 
-/*
 #[inline(never)]
-fn process_memx_memchr_libc(texts: &[&str], pat_byte: u8) -> usize {
-    let mut found: usize = 0;
-    for line in texts {
-        let line_bytes = line.as_bytes();
-        let line_len = line_bytes.len();
-        let mut curr_idx = 0;
-        while curr_idx < line_len {
-            let r = memx::libc::memchr_libc(&line_bytes[curr_idx..], pat_byte);
-            if let Some(pos) = r {
-                found += 1;
-                curr_idx = pos + curr_idx + 1;
-            } else {
-                break;
-            }
-        }
+fn cache_flush(texts: &[&str]) {
+    for i in 0..texts.len() {
+        cache_line_flush(texts[i].as_bytes());
     }
-    found
 }
-*/
 
 mod create_data;
 
@@ -183,55 +168,45 @@ fn criterion_benchmark(c: &mut Criterion) {
         let n = process_memx_memchr_sse2(black_box(&vv), black_box(pat_byte));
         assert_eq!(n, match_cnt);
     }
-    //let n = process_memx_memchr_libc(black_box(&vv), black_box(pat_byte));
-    //assert_eq!(n, match_cnt);
-    memory_barrier(&vv);
+    cache_flush(&vv);
     //
     c.bench_function("std_memchr", |b| {
         b.iter(|| {
             let _r = process_std_memchr(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
     c.bench_function("libc_memchr", |b| {
         b.iter(|| {
             let _r = process_libc_memchr(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
     c.bench_function("memchr_memchr", |b| {
         b.iter(|| {
             let _r = process_memchr_memchr(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
     c.bench_function("memx_memchr", |b| {
         b.iter(|| {
             let _r = process_memx_memchr(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
     c.bench_function("memx_memchr_basic", |b| {
         b.iter(|| {
             let _r = process_memx_memchr_basic(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
     #[cfg(all(any(target_arch = "x86_64", target_arch = "x86"), target_feature = "sse2"))]
     c.bench_function("memx_memchr_sse2", |b| {
         b.iter(|| {
             let _r = process_memx_memchr_sse2(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
+            cache_flush(&vv);
         })
     });
-    /*
-    c.bench_function("memx_memchr_libc", |b| {
-        b.iter(|| {
-            let _r = process_memx_memchr_libc(black_box(&vv), black_box(pat_byte));
-            memory_barrier(&vv);
-        })
-    });
-    */
 }
 
 criterion_group! {