+++ benches/bench-memcpy.rs
@@ -2,6 +2,7 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 mod barrier;
 use barrier::memory_barrier;
+use barrier::cache_line_flush;
 
 #[inline(never)]
 fn process_std_memcpy(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
@@ -9,8 +10,8 @@ fn process_std_memcpy(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
     for i in 0..texts.len() {
         let line_bytes = &mut texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            line_bytes[i..(i + pat_len)].copy_from_slice(pat_bytes);
+        for j in 0..=(line_len - pat_len) {
+            line_bytes[j..(j + pat_len)].copy_from_slice(pat_bytes);
         }
     }
 }
@@ -38,8 +39,8 @@ fn process_libc_memcpy(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
     for i in 0..texts.len() {
         let line_bytes = &mut texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            let _r = _x_libc_memcpy(&mut line_bytes[i..], pat_bytes);
+        for j in 0..=(line_len - pat_len) {
+            let _r = _x_libc_memcpy(&mut line_bytes[j..], pat_bytes);
         }
     }
 }
@@ -50,8 +51,8 @@ fn process_memx_memcpy(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
     for i in 0..texts.len() {
         let line_bytes = &mut texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            let _ = memx::memcpy(&mut line_bytes[i..], pat_bytes);
+        for j in 0..=(line_len - pat_len) {
+            let _ = memx::memcpy(&mut line_bytes[j..], pat_bytes);
         }
     }
 }
@@ -62,8 +63,8 @@ fn process_memx_memcpy_basic(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
     for i in 0..texts.len() {
         let line_bytes = &mut texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            let _ = memx::mem::memcpy_basic(&mut line_bytes[i..], pat_bytes);
+        for j in 0..=(line_len - pat_len) {
+            let _ = memx::mem::memcpy_basic(&mut line_bytes[j..], pat_bytes);
         }
     }
 }
@@ -75,79 +76,104 @@ fn process_memx_memcpy_sse2(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
     for i in 0..texts.len() {
         let line_bytes = &mut texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            let _ = unsafe { memx::arch::x86::_memcpy_sse2(&mut line_bytes[i..], pat_bytes) };
+        for j in 0..=(line_len - pat_len) {
+            let _ = unsafe { memx::arch::x86::_memcpy_sse2(&mut line_bytes[j..], pat_bytes) };
         }
     }
 }
 
-/*
 #[inline(never)]
-fn process_memx_memcpy_libc(texts: &mut [Vec<u8>], pat_bytes: &[u8]) {
+fn assert_result(texts: &[Vec<u8>], pat_bytes: &[u8]) {
     let pat_len = pat_bytes.len();
     for i in 0..texts.len() {
-        let line_bytes = &mut texts[i];
+        let line_bytes = &texts[i];
         let line_len = line_bytes.len();
-        for i in 0..(line_len - pat_len) {
-            let _ = memx::libc::memcpy_libc(&mut line_bytes[i..], pat_bytes);
+        for j in 0..=(line_len - pat_len) {
+            assert_eq!(line_bytes[j], pat_bytes[0]);
+            assert_eq!(&line_bytes[(line_bytes.len() - pat_bytes.len())..], pat_bytes);
         }
     }
 }
-*/
+
+#[inline(never)]
+fn cache_flush(texts: &[Vec<u8>], pattern: &[u8]) {
+    for i in 0..texts.len() {
+        cache_line_flush(&texts[i]);
+    }
+    cache_line_flush(pattern)
+}
 
 mod create_data;
 
 fn criterion_benchmark(c: &mut Criterion) {
     let (mut v, pat_bytes) = create_data::create_data_cpy();
     //
-    process_std_memcpy(black_box(&mut v), black_box(pat_bytes));
-    process_libc_memcpy(black_box(&mut v), black_box(pat_bytes));
-    process_memx_memcpy(black_box(&mut v), black_box(pat_bytes));
-    process_memx_memcpy_basic(black_box(&mut v), black_box(pat_bytes));
+    {
+        let mut v1 = v.clone();
+        process_std_memcpy(black_box(&mut v1), black_box(pat_bytes));
+        assert_result(&v1, pat_bytes);
+    }
+    {
+        let mut v1 = v.clone();
+        process_libc_memcpy(black_box(&mut v1), black_box(pat_bytes));
+        assert_result(&v1, pat_bytes);
+    }
+    {
+        let mut v1 = v.clone();
+        process_memx_memcpy(black_box(&mut v1), black_box(pat_bytes));
+        assert_result(&v1, pat_bytes);
+    }
+    {
+        let mut v1 = v.clone();
+        process_memx_memcpy_basic(black_box(&mut v1), black_box(pat_bytes));
+        assert_result(&v1, pat_bytes);
+    }
     #[cfg(all(any(target_arch = "x86_64", target_arch = "x86"), target_feature = "sse2"))]
-    process_memx_memcpy_sse2(black_box(&mut v), black_box(pat_bytes));
-    //process_memx_memcpy_libc(black_box(&mut v), black_box(pat_bytes));
-    memory_barrier(&mut v);
+    {
+        let mut v1 = v.clone();
+        process_memx_memcpy_sse2(black_box(&mut v1), black_box(pat_bytes));
+        assert_result(&v1, pat_bytes);
+    }
+    cache_flush(&v, pat_bytes);
     //
     c.bench_function("std_memcpy", |b| {
         b.iter(|| {
             process_std_memcpy(black_box(&mut v), black_box(pat_bytes));
             memory_barrier(&mut v);
+            cache_flush(&v, pat_bytes);
         })
     });
     c.bench_function("libc_memcpy", |b| {
         b.iter(|| {
             process_libc_memcpy(black_box(&mut v), black_box(pat_bytes));
             memory_barrier(&mut v);
+            cache_flush(&v, pat_bytes);
         })
     });
     c.bench_function("memx_memcpy", |b| {
         b.iter(|| {
             process_memx_memcpy(black_box(&mut v), black_box(pat_bytes));
             memory_barrier(&mut v);
+            cache_flush(&v, pat_bytes);
         })
     });
     c.bench_function("memx_memcpy_basic", |b| {
         b.iter(|| {
             process_memx_memcpy_basic(black_box(&mut v), black_box(pat_bytes));
             memory_barrier(&mut v);
+            cache_flush(&v, pat_bytes);
         })
     });
     #[cfg(all(any(target_arch = "x86_64", target_arch = "x86"), target_feature = "sse2"))]
-    c.bench_function("memx_memcpy_sse2", |b| {
-        b.iter(|| {
-            process_memx_memcpy_sse2(black_box(&mut v), black_box(pat_bytes));
-            memory_barrier(&mut v);
-        })
-    });
-    /*
-    c.bench_function("memx_memcpy_libc", |b| {
-        b.iter(|| {
-            process_memx_memcpy_libc(black_box(&mut v), black_box(pat_bytes));
-            memory_barrier(&mut v);
-        })
-    });
-    */
+    {
+        c.bench_function("memx_memcpy_sse2", |b| {
+            b.iter(|| {
+                process_memx_memcpy_sse2(black_box(&mut v), black_box(pat_bytes));
+                memory_barrier(&mut v);
+                cache_flush(&v, pat_bytes);
+            })
+        });
+    }
 }
 
 criterion_group! {