+++ src/arch/x86/mod.rs
@@ -71,13 +71,13 @@ pub use x86_cpy::_memcpy_sse2;
 //pub use x86_eq::_memeq_sse2;
 
 //pub use x86_mem::_memmem_avx;
-//pub use x86_mem::_memmem_sse2;
+pub use x86_mem::_memmem_sse2;
 
 pub use x86_rchr::_memrchr_avx;
 pub use x86_rchr::_memrchr_sse2;
 
 //pub use x86_rmem::_memrmem_avx;
-//pub use x86_rmrm::_memrmem_sse2;
+pub use x86_rmem::_memrmem_sse2;
 
 pub use x86_set::_memset_avx;
 pub use x86_set::_memset_sse2;