+++ src/mem/mem_cpy.rs
@@ -153,35 +153,8 @@ pub(crate) fn _memcpy_remaining_15_bytes_impl(
             b_ptr = unsafe { b_ptr.add(loop_size) };
         }
     }
-    {
-        let loop_size = 4;
-        let end_ptr_4 = unsafe { end_ptr.sub(loop_size) };
-        if b_ptr <= end_ptr_4 {
-            _unroll_one_cpy_4!(a_ptr, b_ptr, loop_size, 0);
-            //
-            a_ptr = unsafe { a_ptr.add(loop_size) };
-            b_ptr = unsafe { b_ptr.add(loop_size) };
-        }
-    }
-    {
-        let loop_size = 2;
-        let end_ptr_2 = unsafe { end_ptr.sub(loop_size) };
-        if b_ptr <= end_ptr_2 {
-            _unroll_one_cpy_2!(a_ptr, b_ptr, loop_size, 0);
-            //
-            a_ptr = unsafe { a_ptr.add(loop_size) };
-            b_ptr = unsafe { b_ptr.add(loop_size) };
-        }
-    }
-    {
-        let loop_size = 1;
-        let end_ptr_1 = unsafe { end_ptr.sub(loop_size) };
-        if b_ptr <= end_ptr_1 {
-            _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
-        }
-    }
-    //
-    Ok(())
+    // the remaining data is the max: 7 bytes.
+    _memcpy_remaining_7_bytes_impl(a_ptr, b_ptr, end_ptr)
 }
 
 #[cfg(target_pointer_width = "64")]
@@ -196,6 +169,7 @@ fn _start_cpy_64(dst: &mut [u8], src: &[u8]) -> Result<(), RangeError> {
     let mut b_ptr = src.as_ptr();
     let end_ptr = unsafe { b_ptr.add(src_len) };
     //
+    /*
     {
         let unroll = 8;
         let loop_size = 8;
@@ -209,6 +183,7 @@ fn _start_cpy_64(dst: &mut [u8], src: &[u8]) -> Result<(), RangeError> {
             b_ptr = unsafe { b_ptr.add(loop_size * unroll) };
         }
     }
+    */
     {
         let loop_size = 8;
         let end_ptr_8 = unsafe { end_ptr.sub(loop_size) };
@@ -240,25 +215,8 @@ pub(crate) fn _memcpy_remaining_7_bytes_impl(
             b_ptr = unsafe { b_ptr.add(loop_size) };
         }
     }
-    {
-        let loop_size = 2;
-        let end_ptr_2 = unsafe { end_ptr.sub(loop_size) };
-        if b_ptr <= end_ptr_2 {
-            _unroll_one_cpy_2!(a_ptr, b_ptr, loop_size, 0);
-            //
-            a_ptr = unsafe { a_ptr.add(loop_size) };
-            b_ptr = unsafe { b_ptr.add(loop_size) };
-        }
-    }
-    {
-        let loop_size = 1;
-        let end_ptr_1 = unsafe { end_ptr.sub(loop_size) };
-        if b_ptr <= end_ptr_1 {
-            _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
-        }
-    }
-    //
-    Ok(())
+    // the remaining data is the max: 3 bytes.
+    _memcpy_remaining_3_bytes_impl(a_ptr, b_ptr, end_ptr)
 }
 
 #[cfg(target_pointer_width = "32")]
@@ -307,6 +265,7 @@ pub(crate) fn _memcpy_remaining_3_bytes_impl(
 ) -> Result<(), RangeError> {
     let mut a_ptr = dst_ptr;
     let mut b_ptr = src_ptr;
+    /*
     {
         let loop_size = 2;
         let end_ptr_2 = unsafe { end_ptr.sub(loop_size) };
@@ -321,6 +280,26 @@ pub(crate) fn _memcpy_remaining_3_bytes_impl(
         let loop_size = 1;
         let end_ptr_1 = unsafe { end_ptr.sub(loop_size) };
         if b_ptr <= end_ptr_1 {
+            _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
+        }
+    }
+    */
+    // the remaining 3 bytes data:
+    {
+        let loop_size = 1;
+        if b_ptr < end_ptr {
+            _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
+            //
+            a_ptr = unsafe { a_ptr.add(loop_size) };
+            b_ptr = unsafe { b_ptr.add(loop_size) };
+        }
+        if b_ptr < end_ptr {
+            _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
+            //
+            a_ptr = unsafe { a_ptr.add(loop_size) };
+            b_ptr = unsafe { b_ptr.add(loop_size) };
+        }
+        if b_ptr < end_ptr {
             _unroll_one_cpy_1!(a_ptr, b_ptr, loop_size, 0);
         }
     }
@@ -416,66 +395,6 @@ fn _start_cpy_32_no_unroll(dst: &mut [u8], src: &[u8]) -> Result<(), RangeError>
     // the remaining data is the max: 3 bytes.
     _memcpy_remaining_3_bytes_impl(a_ptr, b_ptr, end_ptr)
 }
-
-/*
-#[inline(always)]
-pub fn _memcpy_impl(dst: &mut [u8], src: &[u8]) -> Result<(), RangeError> {
-    let dst_len = dst.len();
-    let src_len = src.len();
-    if dst_len < src_len {
-        return Err(RangeError);
-    }
-    let mut dst_ptr = dst.as_mut_ptr();
-    let mut src_ptr = src.as_ptr();
-    let end_ptr = unsafe { dst_ptr.add(src_len) };
-    //
-    #[cfg(target_pointer_width = "64")]
-    {
-        let loop_size = 8;
-        while dst_ptr <= unsafe { end_ptr.sub(loop_size) } {
-            let dst_ptr_c8 = dst_ptr as *mut u64;
-            let src_ptr_c8 = src_ptr as *const u64;
-            unsafe {
-                *dst_ptr_c8 = *src_ptr_c8;
-            }
-            dst_ptr = unsafe { dst_ptr.add(loop_size) };
-            src_ptr = unsafe { src_ptr.add(loop_size) };
-        }
-    }
-    //
-    let loop_size = 4;
-    while dst_ptr <= unsafe { end_ptr.sub(loop_size) } {
-        let dst_ptr_c4 = dst_ptr as *mut u32;
-        let src_ptr_c4 = src_ptr as *const u32;
-        unsafe {
-            *dst_ptr_c4 = *src_ptr_c4;
-        }
-        dst_ptr = unsafe { dst_ptr.add(loop_size) };
-        src_ptr = unsafe { src_ptr.add(loop_size) };
-    }
-    /*
-    let loop_size = 2;
-    while dst_ptr <= unsafe { end_ptr.sub(loop_size) } {
-        let dst_ptr_c2 = dst_ptr as *mut u16;
-        let src_ptr_c2 = src_ptr as *const u16;
-        unsafe {
-            *dst_ptr_c2 = *src_ptr_c2;
-        }
-        dst_ptr = unsafe { dst_ptr.add(loop_size) };
-        src_ptr = unsafe { src_ptr.add(loop_size) };
-    }
-    */
-    //
-    while dst_ptr < end_ptr {
-        unsafe {
-            *dst_ptr = *src_ptr;
-        }
-        dst_ptr = unsafe { dst_ptr.add(1) };
-        src_ptr = unsafe { src_ptr.add(1) };
-    }
-    Ok(())
-}
-*/
 
 /*
  * The simple implement: